nyc-taxi-duration-prediction-self-healing-ml/requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=12.0
fastparquet
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
mlflow>=2.8
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.0
shap>=0.43
evidently>=0.4
scipy>=1.11
schedule>=1.2
requests>=2.31
pyyaml>=6.0
ipykernel
pytest>=7.4
httpx>=0.25
